chore(insights): trim duplicate snapshot stories and fold scene stories into chart story files
Removes snapshots that re-covered the same behavior: lifecycle unstacked legend/percentages twins (stacking only changes bar geometry), quill Legend's standalone 4-position set (the BuiltInToggle set pins the same ChartLegendLayout), TimeSeriesBarChart's goal-line/value-label copies of BarChart and ComboChart stories, and the TrendsBarChart BarValueBreakdown story already covered by BarValue50Breakdowns.

The remaining edit-mode scene stories move out of the standalone Funnels/UserPaths/TrendsValue/TrendsLine scene story files into the chart story file for their viz sub-type (as EditScene variants), so each chart's component and scene coverage lives in one place. Shared scene-story parameters and MSW mocks now export from createInsightScene.

Generated-By: PostHog Code
Task-Id: 482e7bd9-6d6b-4b2c-8dfa-a65ba43b5a50
