+++ frontend/src/scenes/insights/__mocks__/createInsightScene.tsx
@@ -1,3 +1,5 @@
+import { samplePersonProperties, sampleRetentionPeopleResponse } from 'scenes/insights/__mocks__/insight.mocks'
+
 import { StoryFn } from '@storybook/react'
 import { useMountedLogic } from 'kea'
 import { router } from 'kea-router'
@@ -6,10 +8,37 @@ import { useOnMountEffect } from 'lib/hooks/useOnMountEffect'
 import { App } from 'scenes/App'
 
 import { sceneLayoutLogic } from '~/layout/scenes/sceneLayoutLogic'
-import { useStorybookMocks } from '~/mocks/browser'
+import { mswDecorator, useStorybookMocks } from '~/mocks/browser'
 import { InsightVizNode, Node } from '~/queries/schema/schema-general'
 import { isInsightVizNode, isLifecycleQuery, isStickinessQuery, isTrendsQuery } from '~/queries/utils'
 import { QueryBasedInsightModel } from '~/types'
+
+/** Spread into a `createInsightStory` story's `parameters`, merging `testOptions` for extra keys. */
+export const insightSceneStoryParameters = {
+    layout: 'fullscreen',
+    viewMode: 'story',
+    mockDate: '2022-03-11',
+    testOptions: {
+        snapshotBrowsers: ['chromium' as const],
+        viewport: {
+            // needs a slightly larger width to push the rendered scene away from the breakpoint boundary
+            width: 1300,
+            height: 720,
+        },
+    },
+}
+
+/** API mocks the insight scene needs beyond the insight itself (editor taxonomy, persons). */
+export const insightSceneMswDecorator = mswDecorator({
+    get: {
+        '/api/environments/:team_id/persons/retention': sampleRetentionPeopleResponse,
+        '/api/environments/:team_id/persons/properties': samplePersonProperties,
+        '/api/projects/:team_id/groups_types': [],
+    },
+    post: {
+        '/api/projects/:team_id/cohorts/': { id: 1 },
+    },
+})
 
 function setLegendFilter(query: Node | null | undefined, showLegend: boolean): Node | null | undefined {
     if (!isInsightVizNode(query)) {
