+++ products/product_analytics/frontend/insights/funnels/FunnelStepsBarChart/FunnelStepsBarChart.stories.tsx
@@ -1,10 +1,19 @@
 import { Meta, StoryObj } from '@storybook/react'
+import { waitFor } from '@testing-library/dom'
+import userEvent from '@testing-library/user-event'
 import { BindLogic } from 'kea'
 import { useState } from 'react'
 
 import { FEATURE_FLAGS } from 'lib/constants'
+import {
+    createInsightStory,
+    insightSceneMswDecorator,
+    insightSceneStoryParameters,
+} from 'scenes/insights/__mocks__/createInsightScene'
 import { insightLogic } from 'scenes/insights/insightLogic'
 
+import funnelLeftToRightFixture from '~/mocks/fixtures/api/projects/team_id/insights/funnelLeftToRight.json'
+import funnelLeftToRightWithInlineEventsFixture from '~/mocks/fixtures/api/projects/team_id/insights/funnelLeftToRightWithInlineEvents.json'
 import funnelTopToBottomFixture from '~/mocks/fixtures/api/projects/team_id/insights/funnelTopToBottom.json'
 import funnelTopToBottomBreakdownFixture from '~/mocks/fixtures/api/projects/team_id/insights/funnelTopToBottomBreakdown.json'
 import funnelTopToBottomBreakdownCompareFixture from '~/mocks/fixtures/api/projects/team_id/insights/funnelTopToBottomBreakdownCompare.json'
@@ -85,3 +94,63 @@ export const BreakdownAndCompare: Story = {
     render: () => <StoryRender insightFixture={funnelTopToBottomBreakdownCompareFixture} />,
     parameters: { featureFlags: [FEATURE_FLAGS.PRODUCT_ANALYTICS_FUNNELS_COMPARE] },
 }
+
+// Full insight scene in edit mode — the steps funnel editor, and the funnels query kind's full data pipeline
+const sceneEditWaitForSelector = ['[data-attr=funnel-steps-bar-chart] canvas[role="img"]', '.PayGateMini']
+
+const waitForFunnelToStabilize: NonNullable<Story['play']> = async ({ canvasElement }) => {
+    let lastHeight = 0
+    await waitFor(
+        () => {
+            const funnelContainer = canvasElement.querySelector('[data-attr=funnel-steps-bar-chart]')
+            const currentHeight = funnelContainer ? funnelContainer.getBoundingClientRect().height : 0
+            if (currentHeight === 0 || currentHeight !== lastHeight) {
+                lastHeight = currentHeight
+                throw new Error('funnel height not yet stable')
+            }
+        },
+        { timeout: 3000, interval: 200 }
+    )
+}
+
+export const EditScene: Story = createInsightStory(funnelLeftToRightFixture as any, 'edit')
+EditScene.decorators = [insightSceneMswDecorator]
+EditScene.parameters = {
+    ...insightSceneStoryParameters,
+    testOptions: { ...insightSceneStoryParameters.testOptions, waitForSelector: sceneEditWaitForSelector },
+}
+
+export const EditSceneWithInlineEvents: Story = createInsightStory(
+    funnelLeftToRightWithInlineEventsFixture as any,
+    'edit'
+)
+EditSceneWithInlineEvents.decorators = [insightSceneMswDecorator]
+EditSceneWithInlineEvents.parameters = {
+    ...insightSceneStoryParameters,
+    testOptions: { ...insightSceneStoryParameters.testOptions, waitForSelector: sceneEditWaitForSelector },
+}
+EditSceneWithInlineEvents.play = async ({ canvasElement }) => {
+    const expandFiltersButton = await waitFor(
+        () => {
+            const filtersButton = canvasElement.querySelector<HTMLElement>('[data-attr="show-prop-filter-0"]')
+            if (!filtersButton) {
+                throw new Error('Filters button not yet rendered')
+            }
+            return filtersButton
+        },
+        { timeout: 2000 }
+    )
+    await userEvent.click(expandFiltersButton)
+}
+
+export const EditSceneViewports: Story = createInsightStory(funnelLeftToRightFixture as any, 'edit')
+EditSceneViewports.decorators = [insightSceneMswDecorator]
+EditSceneViewports.parameters = {
+    ...insightSceneStoryParameters,
+    testOptions: {
+        ...insightSceneStoryParameters.testOptions,
+        waitForSelector: sceneEditWaitForSelector,
+        viewportWidths: ['medium', 'wide', 'superwide'],
+    },
+}
+EditSceneViewports.play = waitForFunnelToStabilize
