+++ products/product_analytics/frontend/insights/trends/TrendsBarChart/TrendsBarChart.stories.tsx
@@ -2,12 +2,16 @@ import { Meta, StoryObj } from '@storybook/react'
 import { BindLogic } from 'kea'
 import { type CSSProperties, useState } from 'react'
 
+import {
+    createInsightStory,
+    insightSceneMswDecorator,
+    insightSceneStoryParameters,
+} from 'scenes/insights/__mocks__/createInsightScene'
 import { insightLogic } from 'scenes/insights/insightLogic'
 import { TrendInsight } from 'scenes/trends/Trends'
 
 import { mswDecorator } from '~/mocks/browser'
 import trendsValueFixture from '~/mocks/fixtures/api/projects/team_id/insights/trendsValue.json'
-import trendsValueBreakdownFixture from '~/mocks/fixtures/api/projects/team_id/insights/trendsValueBreakdown.json'
 import { dataNodeLogic } from '~/queries/nodes/DataNode/dataNodeLogic'
 import type { DataNodeLogicProps } from '~/queries/nodes/DataNode/dataNodeLogic'
 import { insightVizDataNodeKey } from '~/queries/nodes/InsightViz/InsightViz'
@@ -564,6 +568,24 @@ export const BarValue: Story = {
     render: () => renderTrendInsight(trendsValueFixture),
 }
 
-export const BarValueBreakdown: Story = {
-    render: () => renderTrendInsight(trendsValueBreakdownFixture),
+// Full insight scene in edit mode — the trends editor (shared across display types), aggregated result shape
+export const EditScene: Story = createInsightStory(trendsValueFixture as any, 'edit')
+EditScene.decorators = [insightSceneMswDecorator]
+EditScene.parameters = {
+    ...insightSceneStoryParameters,
+    testOptions: {
+        ...insightSceneStoryParameters.testOptions,
+        waitForSelector: '[data-attr=trend-bar-value-graph] > canvas',
+    },
+}
+
+export const EditSceneViewports: Story = createInsightStory(trendsValueFixture as any, 'edit')
+EditSceneViewports.decorators = [insightSceneMswDecorator]
+EditSceneViewports.parameters = {
+    ...insightSceneStoryParameters,
+    testOptions: {
+        ...insightSceneStoryParameters.testOptions,
+        waitForSelector: '[data-attr=trend-bar-value-graph] > canvas',
+        viewportWidths: ['medium', 'wide', 'superwide'],
+    },
 }
